+++ app/includes/post-types.php
@@ -1,0 +1,9 @@
+<?php
+add_action('init', function () use($container) {
+    // post types
+    if ($container->hasParameter('wordpress.post_types')) {
+        foreach ($container->getParameter('wordpress.post_types', []) as $post_type => $args) {
+            register_post_type($post_type, $args);
+        }
+    }
+});