+++ app/includes/session.php
@@ -1,0 +1,7 @@
+<?php
+add_action('init', function () use($container) {
+    // always start a session
+    if (!session_id()) {
+        session_start();
+    }
+});