+++ app/includes/yoast.php
@@ -1,0 +1,5 @@
+<?php
+// move yoast down
+add_filter('wpseo_metabox_prio', function() {
+    return 'low';
+});