+++ functions.php
@@ -1,6 +1,5 @@
 <?php
 require_once __DIR__.'/app/bootstrap.php';
-require_once __DIR__.'/src/functions.php';
 
 add_filter('timber/context', function($data){
     // logos